_code/pyproject.toml
[project]
name = "engram-r"
version = "0.6.0"
description = "Co-scientist EngramR for Obsidian -- generate-debate-evolve hypothesis architecture"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "pandas>=2.1",
    "numpy>=1.26",
    "pyyaml>=6.0",
    "scipy>=1.12",
    "scikit-learn>=1.4",
    "scikit-posthocs>=0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "ruff>=0.4",
    "black>=24.0",
]
bot = [
    "slack-bolt>=1.18",
    "anthropic>=0.39",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.black]
line-length = 88
target-version = ["py311"]
